Moved the test plugins into the test project instead
